refactor: 이미지 업로드 및 리사이즈 미들웨어 통합 및 이름 변경
