--- src/middleware/ImageMiddleware.js
@@ -1,13 +1,0 @@
-import multer from 'multer';
-
-const storage = multer.diskStorage({
-  destination: (req, file, cb) => {
-    cb(null, 'uploads/');
-  },
-  filename: (req, file, cb) => {
-    cb(null, `${Date.now()}-${file.originalname}`);
-  },
-});
-
-const uploads = multer({ storage: storage });
-export default uploads;
